+++ jobbergate-api/CHANGELOG.md
@@ -3,6 +3,9 @@
 This file keeps track of all notable changes to jobbergate-api
 
 ## Unreleased
+
+
+## 5.4.0a3 -- 2024-11-08
 - Added clone capability to job submissions so they can be resubmitted to the cluster when needed [PENG-1676, ASP-4597]
 
 ## 5.4.0a2 -- 2024-11-06